Two Window Testing
Started experimenting to try and find a way to make two windows run at the same time when launching the game. so far I have determined how to get two different cameras to function in the scene, but not two different windows to open when the game is launched. I think we might want to look into some IPC for some assistance in making this happen.