UI and Window Resizing
I determined how to change the window size of the desktop window via the UISetup script and managed to begin putting in the start of some of our UI elements. Some adjustments will be needed and made as required and continued work on putting elements in and adding controls will come later down the line.