More Menus
Added the Audio Settings Menu and the beginning of the Stats menu according to the Figma. A discussion on bars or icons for stat indication and typography is required to continue the design process of the current menus. Functionality work will begin on Monday alongside a continuation of the remaining UI menus.