Falling Items
Player is no able to click on any of the 5 items to drop one from the UI screen. This will eventually be connected to the Looking Glass where a modeled version of the item will drop in for the HoloPal to use. A text build with this addition is added to the repo to demonstrate this change.