UI Item Dropping Adjustments
Made some slight adjustments to the UI so that whenever a need icon is clicked, the button will disappear in order to prevent overloading the system. The button will return after a small amount of time, that being when the dropped object is deleted and the item is being dropped out of the tree. IPC still works, just not in a build. The test build present in the repo right now is outdated. Will hopefully fix next week.