Title Bar Tweaks
There is now a boolean on the canvas item that will allow us to toggle on and off the title bar on the UI window. It was something that was asked about and I was able to put together in the code. Whether we use it or not is up to us but it is there if we choose to use it.