Stats Functionality
The stats bars are currently not working as I need to do IPC stuff again to get the values that I need to grab from HoloPal and send over to the UI scene. That will be my next step forward and then I will circle back around to having the color of the bar change as the stat lowers in value.