IPC Adjustments
Tried to make some adjustments to the surrounding code to determine how to get the IPC working correctly in both directions. Stuck with an error with a thread being aborted ending the sending of messages from the Looking Glass to the UI. UI to Looking Glass IPC still functions as normal.