IPC Continuation
Put together a setup for the remaining stat bars to be able to adjust them later. Starting working on getting Hunger and Thirst to work with the current setup and am currently in a roadblock. There code still works accordingly with no problems so the dirtiness stats bar should still work.