Clean Up and Stats
Started do some code clean up after struggling and blanking on getting IPC for the play function to send the HoloPal back to wandering in the provided space. Added Play and Chat IPC to the stats bars to "update" their value. Decay now works but needs to be seriously slowed down. Hunger is the exception to this as it is directly connected to growth and should be reworked at a later time.