Button Adjutments
Made the play button for dropping needs be inactive when in the egg state, as they are not going to be used while the egg exits. Tried to get the buttons to just become inactive after clicking them and reactivating after some time to increase delay between dropping items but will need to be worked on more.