File Saving/Loading
It works now. The file of the saved data is in a bit of a weird spot as it is per user not, per build, which is expected. If anyone needs to be pointed in the direction to the file, please ask and I can help point in the right direction. there is still one small issue as if values of the stats are 0, they get reset to 1 upon loading but it should be a quick fix.