changed IPC to character t signing
The change is toward the bottom of the IPC script and t replaced the star.

I also included an object in the event system within the DMA canvas scene, just to make that event system less confused. I have received no errors from these changes, and it also eliminated tons of warnings. So, progress.